+++ packages/shared_preferences/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Update shared_preferences to 2.2.0.
 * Update shared_preferences_interface to 2.3.0.
 * Add `clearWithParameters` and `getAllWithParameters`.
+* Update `clear` to use `clearWithParameters`
+* Update `getAll` to use `getAllWithParameters`
 * Update integration_test.
 
 ## 2.1.0